+++ packages/express/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vercel/express
 
+## 0.1.119
+
+### Patch Changes
+
+- @vercel/node@5.9.1
+- @vercel/cervel@0.1.36
+
 ## 0.1.118
 
 ### Patch Changes