+++ packages/node/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vercel/node
 
+## 5.9.1
+
+### Patch Changes
+
+- Updated dependencies [17ee736]
+  - @vercel/build-utils@13.36.1
+
 ## 5.9.0
 
 ### Minor Changes